Don’t allow calling `reveal` in release builds
Some structs have a `reveal` method, which deliberately leaks secret
information for debug builds.  There is no legitimate reason to use
them for anything other than debugging, and no uses at all in any of the
other codebases I am working on.

To ensure that no new uses of this method are added, make this method
only available when debug assertions are turned on.  This ensures that
any attempts to uses this method will break release builds, and thus be
quickly discovered.